training-monitor/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "sentinel-training-monitor"
version = "0.1.0"
description = "Silent Data Corruption detection plugin for GPU training loops"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Sentinel Team" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "grpcio>=1.60",
    "pydantic>=2.0",
    "structlog>=23.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
pytorch = ["torch>=2.2"]
jax = ["jax>=0.4.24", "jaxlib>=0.4.24"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel_training"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
